backend/requirements.txt
fastapi
uvicorn[standard]
duckdb
pandas
google-genai
python-dotenv
pydantic
